Fix implicit sign conversion & integer size truncation warnings
xrefresh.c:125:14: warning: implicit conversion loses integer precision:
      'size_t' (aka 'unsigned long') to 'int' [-Wshorten-64-to-32]
    arglen = strlen (arg);
           ~ ^~~~~~~~~~~~
xrefresh.c:126:12: warning: implicit conversion loses integer precision:
      'size_t' (aka 'unsigned long') to 'int' [-Wshorten-64-to-32]
    slen = strlen (s);
         ~ ^~~~~~~~~~
xrefresh.c:132:26: warning: implicit conversion changes signedness: 'int' to
      'size_t' (aka 'unsigned long') [-Wsign-conversion]
    if (strncmp (arg, s, slen) == 0) return (True);
        ~~~~~~~          ^~~~

+++ xrefresh.c
@@ -114,10 +114,10 @@ parse_boolean_option(char *option)
  */
 
 static Bool 
-isabbreviation(const char *arg, char *s, int minslen)
+isabbreviation(const char *arg, char *s, size_t minslen)
 {
-    int arglen;
-    int slen;
+    size_t arglen;
+    size_t slen;
 
     /* exact match */
     if (strcmp (arg, s) == 0) return (True);
